Move V1 extensions into single extensions folder
