+++ FinancialSummaryApi.Tests/V1/Gateways/DynamoDbGatewayTests.cs
@@ -10,6 +10,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using FinancialSummaryApi.V1.Mappings;
 using Xunit;
 
 namespace FinancialSummaryApi.Tests.V1.Gateways
