+++ FinancialSummaryApi.Tests/V1/UseCase/GetStatementListUseCaseTests.cs
@@ -11,6 +11,7 @@ using Moq;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using FinancialSummaryApi.V1.Mappings;
 using Xunit;
 
 namespace FinancialSummaryApi.Tests.V1.UseCase
