+++ FinancialSummaryApi/V1/Extensions/ExceptionExtensions.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace FinancialSummaryApi.V1.Infrastructure
+namespace FinancialSummaryApi.V1.Extensions
 {
     public static class ExceptionExtensions
     {
